@@ -1,6 +1,5 @@
-import 'package:flow/data/money_flow.dart';
+import 'package:flow/data/chart_data.dart';
 import 'package:flow/entity/category.dart';
-import 'package:flow/entity/transaction.dart';
 import 'package:flow/l10n/extensions.dart';
 import 'package:flow/widgets/home/stats/group_pie_chart.dart';
 import 'package:flow/widgets/home/stats/no_data.dart';
@@ -9,17 +8,15 @@ import 'package:go_router/go_router.dart';
 import 'package:moment_dart/moment_dart.dart';
 
 class PieGraphView extends StatelessWidget {
-  final Map<String, MoneyFlow> data;
+  final Map<String, ChartData> data;
   final TimeRange range;
   final void Function() changeMode;
-  final TransactionType type;
 
   const PieGraphView({
     super.key,
     required this.data,
     required this.range,
     required this.changeMode,
-    required this.type,
   });
 
   @override
@@ -33,7 +30,6 @@ class PieGraphView extends StatelessWidget {
     return SingleChildScrollView(
       padding: const EdgeInsets.only(bottom: 96.0, top: 8.0),
       child: GroupPieChart(
-        type: type,
         data: data,
         unresolvedDataTitle: "category.none".t(context),
         onReselect: (key) {
